pyproject.toml
[project]
name = "detection-engineering"
version = "0.1.0"
description = "jaravis agentic ai system"
authors = [
    {name = "Sivaraj"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph (>=0.6.6,<0.7.0)",
    "langgraph-cli[inmem] (>=0.3.8,<0.4.0)",
    "langgraph-api (>=0.3.1,<0.4.0)",
    "langchain-openai (>=0.3.31,<0.4.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "taxii2-client (>=2.3.0,<3.0.0)",
    "stix2 (>=3.0.1,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "attackcti (>=0.5.4,<0.6.0)",
    "mitreattack-python (>=5.1.0,<6.0.0)",
    "otxv2 (>=1.5.12,<2.0.0)",
    "rapidfuzz (>=3.14.1,<4.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
# Tell Poetry where your actual code lives
packages = [
    { include = "agent", from = "src" }
]